+++ docs/Changelog.md
@@ -3,6 +3,8 @@
 ## New since 0.12.0
 
 - Render themes: symbols on lines with billboard, rotate [#743](https://github.com/mapsforge/vtm/pull/743)
+- Threaded system initialization (Android)
+  - `Parameters.THREADED_INIT`
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.13.0)